+++ src/onegov.core/onegov/core/csv.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """ Offers tools to deal with csv (and xls, xlsx) files. """
 
 from __future__ import absolute_import