+++ src/onegov.core/onegov/core/i18n.py
@@ -162,14 +162,8 @@ def get_translation_bound_meta(meta_class, translate):
         def get_translations(self, form):
             default = super(TranslationBoundMeta, self).get_translations(form)
 
-            # WTForms does this weird dance where it wraps translations for
-            # Python 2, but not for Python 3. That's why we have to do this:
-            if hasattr(default, '_fallback'):
-                if not default._fallback:
-                    default.add_fallback(translate)
-            else:
-                if not default.translations._fallback:
-                    default.translations.add_fallback(translate)
+            if not default._fallback:
+                default.add_fallback(translate)
 
             # Cache for reuse in render_field. I'm not sure this is perfectly
             # sane, as WTForms does use it's own caching mechanism with