+++ src/onegov.core/onegov/core/tests/test_jsx.py
@@ -1,4 +1,4 @@
-from onegov.core.compat import StringIO
+from io import StringIO
 from onegov.core.jsx import JsxFilter